+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Enhancements
 * `FnAccess` and `FnNamespace` now implement `Ord` and `PartialOrd`.
 * The `event_handler_map` example is enhanced to prevent shadowing of the state object map.
 * Separation of constants in function calls is removed as its performance benefit is dubious.
+* A function `sleep` is added to block the current thread by a specified number of seconds.
 
 
 Version 1.5.0